backend/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Admin <admin@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
uvicorn = "^0.12.1"
fastapi = "^0.63.0"
python-multipart = "^0.0.5"
email-validator = "^1.0.5"
requests = "^2.23.0"
celery = "^5.0.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
redis = "^3.5.3"
tenacity = "^7.0.0"
pydantic = "^1.8.1"
emails = "^0.6.0"
sentry-sdk = "^1.0.0"
gunicorn = "^20.0.4"
jinja2 = "^2.11.3"
psycopg2-binary = "^2.8.6"
aiofiles = "^0.6.0"
audioread = "^2.1.9"
faiss-cpu = "^1.7.0"
numpy = "^1.19.5"
pandas = "^1.2.3"
tensorflow = "^2.4.1"
watchdog = "^2.0.2"
argh = "^0.26.2"
PyYAML = "^5.4.1"
alembic = "^1.4.2"
sqlalchemy = "^1.3.19"
pytest = "^6.1.0"
python-jose = {extras = ["cryptography"], version = "^3.2.0"}

[tool.poetry.dev-dependencies]
mypy = "^0.770"
black = "^19.10b0"
isort = "^4.3.21"
autoflake = "^1.3.1"
flake8 = "^3.7.9"
pytest = "^6.1.0"
sqlalchemy-stubs = "^0.3"
pytest-cov = "^2.11.1"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

